Format + clippy cleanup across the tree
cargo fmt: 11 files restyled (mostly use-statement reordering and
let-binding wraps that landed during the unsigned-types and
compiler-split work).

cargo clippy --fix: auto-applied 8 lint fixes:
  * manual is_multiple_of -> .is_multiple_of() (5 sites)
  * useless format!() in a single-arg case
  * manual char-class match -> array literal
  * unnecessary same-type cast in vm::store_i32

Crate-root allow for clippy::doc_lazy_continuation and
::doc_overindented_list_items: the codebase uses `+` and `-` at
the start of comment lines as narrative bullets in declarator /
parser / ABI documentation. Rust 1.95's tightened markdown
heuristic flagged 16 of those as malformed lists; the intent is
paragraphs, not lists, so the silence is at the crate level
rather than via per-comment rewrites.

Verified clean on: aarch64-apple-darwin (host),
aarch64-unknown-linux-gnu, x86_64-unknown-linux-gnu,
x86_64-pc-windows-gnu, aarch64-pc-windows-gnullvm. All 348
tests still pass.

+++ src/c5/vm/mod.rs
@@ -435,7 +435,7 @@ impl<H: Host> Vm<H> {
                 let byte_offset = (addr - STACK_BASE) % 8;
                 let shift = byte_offset * 8;
                 let mask = !(0xFFFF_FFFFu64 << shift);
-                let new_val = (word & mask) | (((val as u32 as u64) << shift) as u64);
+                let new_val = (word & mask) | ((val as u32 as u64) << shift);
                 self.stack[idx] = new_val as i64;
                 Ok(())
             } else {
@@ -750,19 +750,35 @@ impl<H: Host> Vm<H> {
                 // Unsigned compares: same operand layout, but
                 // interpret both stack-top and accumulator as u64.
                 Op::Ult => {
-                    a = if (self.load_i64(sp)? as u64) < (a as u64) { 1 } else { 0 };
+                    a = if (self.load_i64(sp)? as u64) < (a as u64) {
+                        1
+                    } else {
+                        0
+                    };
                     sp += 8;
                 }
                 Op::Ugt => {
-                    a = if (self.load_i64(sp)? as u64) > (a as u64) { 1 } else { 0 };
+                    a = if (self.load_i64(sp)? as u64) > (a as u64) {
+                        1
+                    } else {
+                        0
+                    };
                     sp += 8;
                 }
                 Op::Ule => {
-                    a = if (self.load_i64(sp)? as u64) <= (a as u64) { 1 } else { 0 };
+                    a = if (self.load_i64(sp)? as u64) <= (a as u64) {
+                        1
+                    } else {
+                        0
+                    };
                     sp += 8;
                 }
                 Op::Uge => {
-                    a = if (self.load_i64(sp)? as u64) >= (a as u64) { 1 } else { 0 };
+                    a = if (self.load_i64(sp)? as u64) >= (a as u64) {
+                        1
+                    } else {
+                        0
+                    };
                     sp += 8;
                 }
                 Op::Shl => {
